+++ src/ska_low_mccs_pasd/field_station/field_station_component_manager.py
@@ -592,7 +592,7 @@ class FieldStationComponentManager(TaskExecutorComponentManager):
                     f"FieldStation transitioning to {power_state.name} state ...."
                 )
                 self._power_state = power_state
-                self._component_state_callback(power=power_state)
+                self._update_component_state(power=power_state)
 
         if trimmed_smartbox_power_states:
             with self._power_state_lock: