+++ src/ska_low_mccs_pasd/smart_box/smart_box_device.py
@@ -435,10 +435,12 @@ class MccsSmartBox(MccsBaseDevice):
         )
 
         if communication_state != CommunicationStatus.ESTABLISHED:
-            self._component_state_callback(power=PowerState.UNKNOWN)
+            self.component_manager._update_component_state(power=PowerState.UNKNOWN)
             self._health_monitor_points = {}
         if communication_state == CommunicationStatus.ESTABLISHED:
-            self._component_state_callback(power=self.component_manager._power_state)
+            self.component_manager._update_component_state(
+                power=self.component_manager._power_state
+            )
 
         super()._communication_state_changed(communication_state)