+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Version History
 
 ## Unreleased
+
 * [THORN-253] Populate attribute alarm configuration from tmdata.
 
 ## 3.4.2
+
 * [THORN-305] Prevent unhandled exception when handling errors from write requests.
 * [SKB-455] Re-create the Modbus client after a failed poll to workaround the connection issues.